ali baba

+++ Competitive_programming/alibaba.cpp
@@ -1,7 +1,7 @@
 #include<iostream>
 using namespace std;
 int main(){
-    int a,b,c,d;
+    long long a,b,c,d;
     cin>>a>>b>>c>>d;
     if((a+b-c==d)|| (a+(b*c)==d)|| (a-b+c==d)|| (a-(b*c)==d) || ((a*b)+c==d) || ((a*b)-c==d)){
         cout<<"YES";
